feat: Add SortColors.py for sorting an array of colors
This commit adds a new file, SortColors.py, which contains two functions for sorting an array of colors. The first function, sort_colors, implements the Dutch National Flag Algorithm to sort the colors in-place. The second function, sort_colors2, uses a counter approach to sort the colors.

The SortColors.py file also includes test cases for both functions.

The addition of this file enhances the functionality of the repository by providing a solution for sorting arrays of colors.